requirements.txt
#pymysql
pymysql
#flask-sqlalchemy
flask-sqlalchemy
#flask-script
flask-script
#flask-migrate
flask-migrate